More crypto refactoring
